+++ src/node/utils/tar.json
@@ -68,7 +68,8 @@
   , "security.js"
   , "$security.js"
   , "json2.js"
-  , "pluginfw/plugins.js"
+  , "pluginfw/client_plugins.js"
+  , "pluginfw/shared.js"
   , "pluginfw/hooks.js"
   , "pluginfw/parent_require.js"
   ]
