+++ src/static/js/ace.js
@@ -244,7 +244,7 @@ require.setGlobalKeyPath("require");\n\
 <script type="text/javascript">\
   parent_req = require("ep_etherpad-lite/static/js/pluginfw/parent_require");\
   parent_req.getRequirementFromParent(require, "ep_etherpad-lite/static/js/pluginfw/hooks");\
-  parent_req.getRequirementFromParent(require, "ep_etherpad-lite/static/js/pluginfw/plugins");\
+  parent_req.getRequirementFromParent(require, "ep_etherpad-lite/static/js/pluginfw/client_plugins");\
 </script>\
 ');
