+++ src/static/js/pluginfw/plugins.js
@@ -1,30 +1,21 @@
-exports.isClient = typeof global != "object";
+var npm = require("npm/lib/npm.js");
+var readInstalled = require("./read-installed.js");
+var relativize = require("npm/lib/utils/relativize.js");
+var readJson = require("npm/lib/utils/read-json.js");
+var path = require("path");
+var async = require("async");
+var fs = require("fs");
+var tsort = require("./tsort");
+var util = require("util");
+var _ = require("underscore");
 
-var _;
-
-if (!exports.isClient) {
-  var npm = require("npm/lib/npm.js");
-  var readInstalled = require("./read-installed.js");
-  var relativize = require("npm/lib/utils/relativize.js");
-  var readJson = require("npm/lib/utils/read-json.js");
-  var path = require("path");
-  var async = require("async");
-  var fs = require("fs");
-  var tsort = require("./tsort");
-  var util = require("util");
-  _ = require("underscore");
-}else{
-  var $, jQuery;
-  $ = jQuery = require("ep_etherpad-lite/static/js/rjquery").$;
-  _ = require("ep_etherpad-lite/static/js/underscore");
-}
+var pluginUtils = require('./shared');
 
 exports.prefix = 'ep_';
 exports.loaded = false;
 exports.plugins = {};
 exports.parts = [];
 exports.hooks = {};
-exports.baseURL = '';
 
 exports.ensure = function (cb) {
   if (!exports.loaded)
@@ -43,7 +34,7 @@ exports.formatParts = function () {
 
 exports.formatHooks = function (hook_set_name) {
   var res = [];
-  var hooks = exports.extractHooks(exports.parts, hook_set_name || "hooks");
+  var hooks = pluginUtils.extractHooks(exports.parts, hook_set_name || "hooks");
 
   _.chain(hooks).keys().forEach(function (hook_name) {
     _.forEach(hooks[hook_name], function (hook) {
@@ -52,84 +43,6 @@ exports.formatHooks = function (hook_set_name) {
   });
   return "<dl>" + res.join("\n") + "</dl>";
 };
-
-function loadFn(path, hookName) {
-  var functionName
-    , parts = path.split(":");
-  
-  // on windows: C:\foo\bar:xyz
-  if(parts[0].length == 1) {
-    if(parts.length == 3)
-      functionName = parts.pop();
-    path = parts.join(":");
-  }else{
-    path = parts[0];
-    functionName = parts[1];
-  }
-  
-  var fn = require(path);
-  functionName = functionName ? functionName : hookName;  
-
-  _.each(functionName.split("."), function (name) {
-    fn = fn[name];
-  });
-  return fn;
-};
-
-function extractHooks(parts, hook_set_name, normalizer) {
-  var hooks = {};
-  _.each(parts,function (part) {
-    _.chain(part[hook_set_name] || {})
-    .keys()
-    .each(function (hook_name) {
-      if (hooks[hook_name] === undefined) hooks[hook_name] = [];
-      
-      var hook_fn_name = part[hook_set_name][hook_name];
-
-      /* On the server side, you can't just
-       * require("pluginname/whatever") if the plugin is installed as
-       * a dependency of another plugin! Bah, pesky little details of
-       * npm... */
-      if (normalizer) {
-        hook_fn_name = normalizer(part, hook_fn_name);
-      }
-
-      try {
-        var hook_fn = loadFn(hook_fn_name, hook_name);
-        if (!hook_fn) {
-          throw "Not a function";
-        }
-      } catch (exc) {
-        console.error("Failed to load '" + hook_fn_name + "' for '" + part.full_name + "/" + hook_set_name + "/" + hook_name + "': " + exc.toString())
-      }
-      if (hook_fn) {
-        hooks[hook_name].push({"hook_name": hook_name, "hook_fn": hook_fn, "hook_fn_name": hook_fn_name, "part": part});
-      }
-    });
-  });
-  return hooks;
-};
-
-
-if (exports.isClient) {
-  exports.update = function (cb) {
-    // It appears that this response (see #620) may interrupt the current thread
-    // of execution on Firefox. This schedules the response in the run-loop,
-    // which appears to fix the issue.
-    var callback = function () {setTimeout(cb, 0);};
-
-    jQuery.getJSON(exports.baseURL + 'pluginfw/plugin-definitions.json', function(data) {
-      exports.plugins = data.plugins;
-      exports.parts = data.parts;
-      exports.hooks = extractHooks(exports.parts, "client_hooks");
-      exports.loaded = true;
-      callback();
-     }).error(function(xhr, s, err){
-       console.error("Failed to load plugin-definitions: " + err);
-       callback();
-     });
-  };
-} else {
 
 exports.callInit = function (cb) {
   var hooks = require("./hooks");
@@ -171,7 +84,7 @@ exports.update = function (cb) {
         if (err) cb(err);
         exports.plugins = plugins;
         exports.parts = sortParts(parts);
-        exports.hooks = exports.extractHooks(exports.parts, "hooks", exports.pathNormalization);
+        exports.hooks = pluginUtils.extractHooks(exports.parts, "hooks", exports.pathNormalization);
         exports.loaded = true;
         exports.callInit(cb);
       }
@@ -205,9 +118,6 @@ exports.getPackages = function (cb) {
     cb(null, packages);
   });
 };
-
-}
-
 
 function loadPlugin(packages, plugin_name, plugins, parts, cb) {
   var plugin_path = path.resolve(packages[plugin_name].path, "ep.json");
