+++ src/templates/pad.html
@@ -369,7 +369,7 @@
                 document.domain = document.domain; // for comet
               }
 
-              var plugins = require('ep_etherpad-lite/static/js/pluginfw/plugins');
+              var plugins = require('ep_etherpad-lite/static/js/pluginfw/client_plugins');
               var hooks = require('ep_etherpad-lite/static/js/pluginfw/hooks');
 
               plugins.baseURL = baseURL;
