+++ src/templates/timeslider.html
@@ -175,7 +175,7 @@
       document.domain = document.domain; // for comet
     }
 
-    var plugins = require('ep_etherpad-lite/static/js/pluginfw/plugins');
+    var plugins = require('ep_etherpad-lite/static/js/pluginfw/client_plugins');
     plugins.baseURL = baseURL;
     
     plugins.update(function () {
